+++ src/thor/multimodal.cc
@@ -51,8 +51,7 @@ MultiModalPathAlgorithm::~MultiModalPathAlgorithm() {
 }
 
 // Initialize prior to finding best path
-void MultiModalPathAlgorithm::Init(const midgard::PointLL& destll,
-                                   const std::shared_ptr<DynamicCost>& costing) {
+void MultiModalPathAlgorithm::Init(const midgard::PointLL& destll, const cost_ptr_t& costing) {
   // Disable A* for multimodal
   astarheuristic_.Init(destll, 0.0f);
 
@@ -229,8 +228,8 @@ bool MultiModalPathAlgorithm::ExpandForward(GraphReader& graphreader,
                                             const MMEdgeLabel& pred,
                                             const uint32_t pred_idx,
                                             const bool from_transition,
-                                            const std::shared_ptr<DynamicCost>& pc,
-                                            const std::shared_ptr<DynamicCost>& tc,
+                                            const cost_ptr_t& pc,
+                                            const cost_ptr_t& tc,
                                             const sif::mode_costing_t& mode_costing,
                                             const TimeInfo& time_info) {
 
@@ -544,7 +543,7 @@ bool MultiModalPathAlgorithm::ExpandForward(GraphReader& graphreader,
 void MultiModalPathAlgorithm::SetOrigin(GraphReader& graphreader,
                                         valhalla::Location& origin,
                                         const valhalla::Location& destination,
-                                        const std::shared_ptr<DynamicCost>& costing) {
+                                        const cost_ptr_t& costing) {
   // Only skip inbound edges if we have other options
   bool has_other_edges = false;
   std::for_each(origin.correlation().edges().begin(), origin.correlation().edges().end(),
@@ -652,7 +651,7 @@ void MultiModalPathAlgorithm::SetOrigin(GraphReader& graphreader,
 // Add a destination edge
 uint32_t MultiModalPathAlgorithm::SetDestination(GraphReader& graphreader,
                                                  const valhalla::Location& dest,
-                                                 const std::shared_ptr<DynamicCost>& costing) {
+                                                 const cost_ptr_t& costing) {
   // Only skip outbound edges if we have other options
   bool has_other_edges =
       std::any_of(dest.correlation().edges().begin(), dest.correlation().edges().end(),
@@ -698,7 +697,7 @@ bool MultiModalPathAlgorithm::ExpandFromNode(baldr::GraphReader& graphreader,
                                              const baldr::GraphId& node,
                                              const sif::EdgeLabel& pred,
                                              const uint32_t pred_idx,
-                                             const std::shared_ptr<DynamicCost>& costing,
+                                             const cost_ptr_t& costing,
                                              EdgeStatus& edgestatus,
                                              std::vector<EdgeLabel>& edgelabels,
                                              DoubleBucketQueue<EdgeLabel>& adjlist,
@@ -779,7 +778,7 @@ bool MultiModalPathAlgorithm::ExpandFromNode(baldr::GraphReader& graphreader,
 bool MultiModalPathAlgorithm::CanReachDestination(const valhalla::Location& destination,
                                                   GraphReader& graphreader,
                                                   const travel_mode_t dest_mode,
-                                                  const std::shared_ptr<DynamicCost>& costing) {
+                                                  const cost_ptr_t& costing) {
   // Assume pedestrian mode for now
   mode_ = dest_mode;